CTFd/plugins/custom/requirements.txt
Flask-Cors==3.0.10
Flask-Uploads==0.2.1
python_slugify==5.0.2
docx2pdf==0.1.8